🧪 Testing Improvement: Extract and refine Mp3PropertyGetter tests
🎯 What: Extracted Mp3PropertyGetter tests from the monolithic PropertyGetterTests.cs into a dedicated Mp3PropertyGetterTests.cs file, improving maintainability and isolation.
📊 Coverage: Covered all original scenarios (happy path, file not found, invalid formats, null handling) using a dedicated Mp3Setup class that dynamically provisions lightweight mock MP3 files via Base64 embedding.
✨ Result: Enhanced test coverage, separation of concerns, and alignment with project testing patterns.
