+++ HDLG.Tests/PropertyGetterTests.cs
@@ -212,88 +212,6 @@ namespace HDLG.Tests
 
 
         [Fact]
-        public void Mp3PropertyGetter_AddLogger_SetsLogger()
-        {
-            // Arrange
-            var getter = new Mp3PropertyGetter();
-
-            // Act
-            getter.AddLogger(loggerMock.Object);
-
-            // Assert
-            getter.Logger.Should().Be(loggerMock.Object);
-        }
-
-        [Fact]
-        public void Mp3PropertyGetter_AddLogger_NullLogger_ThrowsArgumentNullException()
-        {
-            // Arrange
-            var getter = new Mp3PropertyGetter();
-
-            // Act & Assert
-            Assert.Throws<ArgumentNullException>(() => getter.AddLogger(null!));
-        }
-
-        [Fact]
-        public void Mp3PropertyGetter_GetFileProperties_AllProperties_ReturnsCompleteDictionary()
-        {
-            // Arrange
-            var getter = new Mp3PropertyGetter();
-            var testFile = "test_full.mp3";
-
-            // Create a test file with all properties filled to verify they are extracted correctly
-            System.IO.File.Copy("test.mp3", testFile, true);
-            using (var f = TagLib.File.Create(testFile))
-            {
-                f.Tag.Title = "Test Complete Title";
-                f.Tag.Album = "Test Complete Album";
-                f.Tag.Year = 2024;
-                f.Tag.Performers = new[] { "John Doe", "Jane Doe" };
-                f.Tag.AlbumArtists = new[] { "Band A", "Band B" };
-                f.Tag.Composers = new[] { "Mozart", "Beethoven" };
-                f.Tag.Copyright = "2024 Test Corp";
-                f.Save();
-            }
-
-            try
-            {
-                // Act
-                var properties = getter.GetFileProperties(testFile);
-
-                // Assert
-                properties.Should().ContainKey("Title");
-                properties["Title"].Should().Be("Test Complete Title");
-
-                properties.Should().ContainKey("Album");
-                properties["Album"].Should().Be("Test Complete Album");
-
-                properties.Should().ContainKey("Year");
-                properties["Year"].Should().Be(2024u);
-
-                properties.Should().ContainKey("Performers");
-                properties["Performers"].Should().Be("John Doe, Jane Doe");
-
-                properties.Should().ContainKey("AlbumArtists");
-                properties["AlbumArtists"].Should().Be("Band A, Band B");
-
-                properties.Should().ContainKey("Composers");
-                properties["Composers"].Should().Be("Mozart, Beethoven");
-
-                properties.Should().ContainKey("Copyright");
-                properties["Copyright"].Should().Be("2024 Test Corp");
-
-                properties.Should().ContainKey("Duration");
-            }
-            finally
-            {
-                if (System.IO.File.Exists(testFile))
-                {
-                    System.IO.File.Delete(testFile);
-                }
-            }
-        }
-
-        [Fact]
         public void ImagePropertyGetter_GetFileProperties_CorruptedImage_LogsWarningAndReturnsEmpty()
         {
             // Arrange
@@ -307,86 +225,6 @@ namespace HDLG.Tests
             properties.Should().BeEmpty();
             loggerMock.Verify(l => l.Warning(It.IsAny<Exception>(), It.Is<string>(s => s.Contains("Invalid image content")), It.IsAny<string>()), Times.Once);
         }
-        [Theory]
-        [InlineData("test.mp3", true)]
-        [InlineData("test.flac", true)]
-        [InlineData("test.wav", true)]
-        [InlineData("test.txt", false)]
-        [InlineData("test.jpg", false)]
-        public void Mp3PropertyGetter_IsSupportedFile_ReturnsExpectedResult(string path, bool expected)
-        {
-            // Arrange
-            var getter = new Mp3PropertyGetter();
-
-            // Act
-            var result = getter.IsSupportedFile(path);
-
-            // Assert
-            result.Should().Be(expected);
-        }
-
-        [Fact]
-        public void Mp3PropertyGetter_GetFileProperties_ValidFileWithTitle_ReturnsProperties()
-        {
-            // Arrange
-            var getter = new Mp3PropertyGetter();
-
-            // Act
-            var properties = getter.GetFileProperties("test.mp3");
-
-            // Assert
-            properties.Should().ContainKey("Title");
-            properties["Title"].Should().Be("Test Title");
-            properties.Should().ContainKey("Album");
-            properties["Album"].Should().Be("Test Album");
-            properties.Should().ContainKey("Year");
-            properties["Year"].Should().Be(0u);
-        }
-
-        [Fact]
-        public void Mp3PropertyGetter_GetFileProperties_FileNotFound_LogsErrorAndReturnsEmpty()
-        {
-            // Arrange
-            var getter = new Mp3PropertyGetter();
-            getter.AddLogger(loggerMock.Object);
-
-            // Act
-            var properties = getter.GetFileProperties("nonexistent.mp3");
-
-            // Assert
-            loggerMock.Verify(l => l.Error(It.IsAny<Exception>(), It.Is<string>(s => s.Contains("Cannot read file")), It.IsAny<string>()), Times.Once);
-            properties.Should().BeEmpty();
-        }
-
-        [Fact]
-        public void Mp3PropertyGetter_GetFileProperties_InvalidFileFormat_LogsWarningAndReturnsEmpty()
-        {
-            // Arrange
-            var getter = new Mp3PropertyGetter();
-            getter.AddLogger(loggerMock.Object);
-            var invalidFile = "test_invalid.mp3";
-            System.IO.File.WriteAllText(invalidFile, "invalid mp3 content");
-
-            try
-            {
-                // Act
-                var properties = getter.GetFileProperties(invalidFile);
-
-                // Assert
-                loggerMock.Verify(l => l.Warning(It.IsAny<Exception>(), It.Is<string>(s => s.Contains("is corrupted") || s.Contains("is not supported") || s.Contains("Cannot read properties from file")), It.IsAny<string>()), Times.Once);
-                properties.Should().BeEmpty();
-            }
-            finally
-            {
-                if (System.IO.File.Exists(invalidFile))
-                {
-                    System.IO.File.Delete(invalidFile);
-                }
-            }
-        }
-
-
-
         [Fact]
         public void PdfPropertyGetter_GetFileProperties_ValidFileWithTitle_ReturnsTitle()
         {
